made farm contract single token farm by default. kept the old farm separately
