+++ elrond_dex_farm/src/config.rs
@@ -1,0 +1,49 @@
+elrond_wasm::imports!();
+elrond_wasm::derive_imports!();
+
+#[derive(TopEncode, TopDecode, PartialEq, TypeAbi)]
+pub enum State {
+    Inactive,
+    Active,
+}
+
+#[elrond_wasm_derive::module]
+pub trait ConfigModule {
+    #[inline]
+    fn is_active(&self) -> bool {
+        let state = self.state().get();
+        state == State::Active
+    }
+
+    fn require_permissions(&self) -> SCResult<()> {
+        let caller = self.blockchain().get_caller();
+        let owner = self.owner().get();
+        let router = self.router_address().get();
+        require!(caller == owner || caller == router, "Permission denied");
+        Ok(())
+    }
+
+    #[view(getLastErrorMessage)]
+    #[storage_mapper("last_error_message")]
+    fn last_error_message(&self) -> SingleValueMapper<Self::Storage, BoxedBytes>;
+
+    #[view(getRouterAddress)]
+    #[storage_mapper("router_address")]
+    fn router_address(&self) -> SingleValueMapper<Self::Storage, Address>;
+
+    #[view(getState)]
+    #[storage_mapper("state")]
+    fn state(&self) -> SingleValueMapper<Self::Storage, State>;
+
+    #[view(getOwner)]
+    #[storage_mapper("owner")]
+    fn owner(&self) -> SingleValueMapper<Self::Storage, Address>;
+
+    #[view(getFarmingTokenId)]
+    #[storage_mapper("farming_token_id")]
+    fn farming_token_id(&self) -> SingleValueMapper<Self::Storage, TokenIdentifier>;
+
+    #[view(getRewardTokenId)]
+    #[storage_mapper("reward_token_id")]
+    fn reward_token_id(&self) -> SingleValueMapper<Self::Storage, TokenIdentifier>;
+}
