+++ multi_token_farm/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "elrond_dex_farm"
+version = "0.0.0"
+authors = [ "you",]
+edition = "2018"
+publish = false
+
+[lib]
+path = "src/lib.rs"
+
+[features]
+wasm-output-mode = ["elrond-wasm-node"]
+
+[dependencies.elrond-wasm]
+version = "0.16"
+
+[dependencies.elrond-wasm-derive]
+version = "0.16"
+
+[dependencies.elrond-wasm-node]
+version = "0.16"
+optional = true
+
+[dev-dependencies.elrond-wasm-debug]
+version = "0.16"
+
+[dependencies.elrond_dex_pair]
+path = "../elrond_dex_pair"
+
+[dependencies.dex-common]
+path = "../common/dex-common"
