+++ multi_token_farm/abi/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "elrond_dex_farm-abi"
+
+version = "0.0.0"
+authors = [ "you",]
+edition = "2018"
+publish = false
+
+[dependencies.elrond_dex_farm]
+path = ".."
+
+[dependencies.elrond-wasm]
+version = "0.16"
+
+[dependencies.elrond-wasm-debug]
+version = "0.16"
