+++ multi_token_farm/abi/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+	elrond_wasm_debug::abi_json::print_abi::<elrond_dex_farm::AbiProvider>();
+}
