+++ multi_token_farm/elrond.json
@@ -1,0 +1,3 @@
+{
+    "language": "rust"
+}
