+++ multi_token_farm/wasm/src/lib.rs
@@ -1,0 +1,4 @@
+#![no_std]
+
+pub use elrond_dex_farm::*;
+pub use elrond_wasm_output::*;
